+++ routes/webhook.php
@@ -5,12 +5,11 @@ declare(strict_types=1);
 use Gowelle\BeemAfrica\Http\Controllers\CollectionWebhookController;
 use Gowelle\BeemAfrica\Http\Controllers\MojaWebhookController;
 use Gowelle\BeemAfrica\Http\Controllers\SmsWebhookController;
-use Gowelle\BeemAfrica\Http\Controllers\UssdWebhookController;
 use Gowelle\BeemAfrica\Http\Controllers\WebhookController;
 use Illuminate\Support\Facades\Route;
 
-Route::post(config('beem-africa.webhook.path', 'webhooks/beem'), WebhookController::class)
-    ->middleware(config('beem-africa.webhook.middleware', []))
+Route::post(config('beem-africa.checkout.webhook_path', 'webhooks/beem'), WebhookController::class)
+    ->middleware(config('beem-africa.checkout.webhook_middleware', []))
     ->name('beem.webhook');
 
 // SMS Delivery Report Webhook
@@ -30,12 +29,6 @@ Route::post(
     config('beem-africa.collection.webhook_path', 'webhooks/beem/collection'),
     [CollectionWebhookController::class, 'handleCallback']
 )->name('beem.collection');
-
-// USSD Session Callback
-Route::post(
-    config('beem-africa.ussd.webhook_path', 'webhooks/beem/ussd'),
-    [UssdWebhookController::class, 'handleCallback']
-)->name('beem.ussd');
 
 // Moja Incoming Message Webhook
 Route::post(